Revert "extractAllNumbers() - changed return value to double pointer instead of integer pointer"
This reverts commit 27c3ec408941bf8fc0da519b60392e9e50d41950.

+++ win32Calculator/BUTTONS.cpp
@@ -1457,11 +1457,11 @@ bool BUTTONS::CLICKED::Equal()
 	const int iTotalNumbers = mathParser.countNumbers(staticText, staticTextSize, true);
 
 	//Dynamically allocates array for storing all numbers to be calculated
-	double* iWorkingSetNumbers = new double[iTotalNumbers];
+	int* iWorkingSetNumbers = new int[iTotalNumbers];
 
 	iWorkingSetNumbers = mathParser.extractAllNumbers(staticText, staticTextSize);
 
-	double check = -1;
+	int check = -1;
 
 	for (int i = 0; i < iTotalNumbers; i++)
 	{
